requirements.txt
pandas==2.3.3
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.52.2
tldextract==5.3.1
